+++ tests/Eccube/Tests/Twig/Extension/FunctionsTest.php
@@ -14,16 +14,9 @@
 namespace Eccube\Tests\Twig\Extension;
 
 use Eccube\Tests\EccubeTestCase;
-use org\bovigo\vfs\vfsStream;
 
 class FunctionsTest extends EccubeTestCase
 {
-    protected $templateDir;
-    protected $blockTwigs = [
-        'test_block.twig',
-        'test_block2.twig',
-    ];
-
     /**
      * @var \Twig_Environment
      */
@@ -33,31 +26,7 @@ class FunctionsTest extends EccubeTestCase
     {
         parent::setUp();
 
-        $client = self::createClient();
-        $root = vfsStream::setup();
-        $this->templateDir = $root->url();
-
-        // Rewrite to parameters using reflection
-        $container = $client->getContainer();
-        $parameters = $container->getParameterBag()->all();
-        $parameters['eccube_twig_block_templates'] = $this->blockTwigs;
-
-        $refClass = new \ReflectionClass($container);
-        $refProp = $refClass->getProperty('parameters');
-        $refProp->setAccessible(true);
-        $refProp->setValue($container, $parameters);
-
-        // Initialize to the Twig loader
-        $this->twig = $client->getContainer()->get('twig');
-        $paths = $this->twig->getLoader()->getPaths();
-        $paths[] = $this->templateDir;
-        $this->twig->setLoader(new \Twig_Loader_Filesystem($paths));
-        $this->twig->setCache(false);
-
-        foreach ($this->blockTwigs as $twig_file) {
-            // Preventing undefined errors
-            file_put_contents($this->templateDir.'/'.$twig_file, '');
-        }
+        $this->twig = $this->container->get('twig');
     }
 
     /**
@@ -70,44 +39,6 @@ class FunctionsTest extends EccubeTestCase
         $this->expected = "<div id='test'>aaa</div>";
         $this->actual = $template->render([]);
 
-        $this->verify();
-    }
-
-    /**
-     * 任意のブロックをユーザー定義関数で出力するテスト.
-     */
-    public function testEccubeBlockFunctions()
-    {
-        $file = $this->templateDir.'/'.$this->blockTwigs[0];
-        $source = '{% block exampleblock %}<div id="exampleblock">test</div>{% endblock %}';
-
-        file_put_contents($file, $source);
-        $template = $this->twig->createTemplate("<div id='test'>{{ eccube_block_exampleblock() }}</div>");
-
-        $this->expected = "<div id='test'><div id=\"exampleblock\">test</div></div>";
-        $this->actual = $template->render([]);
-
-        $this->verify();
-    }
-
-    /**
-     * 任意のブロックをユーザー定義関数で出力するテスト.
-     *
-     * 引数付き
-     */
-    public function testEccubeBlockFunctionsWithParams()
-    {
-        $file = $this->templateDir.'/'.$this->blockTwigs[1];
-        $source = '{% block exampleblock2 %}<div id="exampleblock">{{ variable }}</div>{% endblock %}';
-
-        file_put_contents($file, $source);
-
-        $template = $this->twig->createTemplate(
-            "<div id='test'>{{ eccube_block_exampleblock2({'variable': 'example'}) }}</div>"
-        );
-
-        $this->expected = "<div id='test'><div id=\"exampleblock\">example</div></div>";
-        $this->actual = $template->render(['variable' => 'example']);
         $this->verify();
     }
 }
